Separate out Timer into its own module
