+++ azure-functions-sdk/src/commands/new.rs
@@ -19,10 +19,12 @@ use syn::{parse::Parser, parse_file, punctuated::Punctuated, Ident, Item, Token}
 mod blob;
 mod http;
 mod queue;
+mod timer;
 
 pub use self::blob::Blob;
 pub use self::http::Http;
 pub use self::queue::Queue;
+pub use self::timer::Timer;
 
 fn get_path_for_function(name: &str) -> Result<String, String> {
     if !Regex::new("^[a-zA-Z][a-zA-Z0-9_]*$")
@@ -211,52 +213,6 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
             quiet: args.is_present("quiet"),
             color: args.value_of("color"),
             args,
-        }
-    }
-}
-
-struct Timer<'a> {
-    name: &'a str,
-    schedule: &'a str,
-}
-
-impl<'a> Timer<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("timer")
-            .about("Creates a new timer triggered Azure Function.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-            .arg(
-                Arg::with_name("schedule")
-                    .long("schedule")
-                    .short("s")
-                    .value_name("SCHEDULE")
-                    .help("The timer schedule as a cron-expression.")
-                    .required(true),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-            "schedule": self.schedule,
-        });
-
-        create_function(self.name, "timer.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for Timer<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        Timer {
-            name: args.value_of("name").unwrap(),
-            schedule: args.value_of("schedule").unwrap(),
         }
     }
 }
